requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-admin-list-filter-dropdown==1.0.3
django-extensions==3.0.5
django-filter==2.3.0
django-heroku==0.3.1
django-seed==0.2.2
django-storages==1.10.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
dropbox==10.6.0
et-xmlfile==1.0.1
Faker==4.9.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
imutils==0.5.3
jdcal==1.4.1
logdna==1.5.3
Markdown==3.2.2
numpy==1.19.3
opencv-python==4.4.0.44
openpyxl==3.0.5
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==1.7.1
pymongo==3.10.1
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
whitenoise==5.2.0
xlrd==1.2.0

zipp==3.1.0

